Increase buffer size for value editing, fix warning with size_t -> int conversion.

+++ src/debugger.cpp
@@ -21,7 +21,7 @@ void Debugger::Remove(std::string name) {
 }
 
 int Debugger::size() {
-	return Debugger::log.size();
+	return (int)Debugger::log.size();
 }
 
 bool Debugger::empty() {
